Fix Test02 rename
