+++ AngularJsTest02/App_Start/FilterConfig.cs
@@ -1,7 +1,7 @@
 ﻿using System.Web;
 using System.Web.Mvc;
 
-namespace AngularJsTest01
+namespace AngularJsTest02
 {
     public class FilterConfig
     {
