+++ AngularJsTest02/App_Start/RouteConfig.cs
@@ -5,7 +5,7 @@ using System.Web;
 using System.Web.Mvc;
 using System.Web.Routing;
 
-namespace AngularJsTest01
+namespace AngularJsTest02
 {
     public class RouteConfig
     {
