+++ AngularJsTest02/Global.asax
@@ -1,1 +1,1 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="AngularJsTest01.MvcApplication" Language="C#" %>
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="AngularJsTest02.MvcApplication" Language="C#" %>
