+++ AngularJsTest02/Global.asax.cs
@@ -7,7 +7,7 @@ using System.Web.Mvc;
 using System.Web.Optimization;
 using System.Web.Routing;
 
-namespace AngularJsTest01
+namespace AngularJsTest02
 {
     public class MvcApplication : System.Web.HttpApplication
     {
