+++ AngularJsTest02/Views/Web.config
@@ -17,7 +17,7 @@
         <add namespace="System.Web.Mvc.Html" />
         <add namespace="System.Web.Optimization"/>
         <add namespace="System.Web.Routing" />
-        <add namespace="AngularJsTest01" />
+        <add namespace="AngularJsTest02" />
       </namespaces>
     </pages>
   </system.web.webPages.razor>
